splits "StatsColumn" into individual components to rework individually
